Fail if ':' is passed to binary_to_integer/2
Before:

1> binary_to_integer(<<":">>, 16).
3

After:

1> binary_to_integer(<<":">>, 16).
** exception error: bad argument
     in function  binary_to_integer/2
        called as binary_to_integer(<<":">>,16)

Prior to this change, both list_to_integer/2 and binary_to_integer/2
would convert strings with values between ASCII '9' up to '0'+base for
base > 10. For example, when converting in base 16, you could pass ':',
';', '<', '=', '>', and '?' without getting an exception.  This was due
to a missing check in c2int_is_invalid_char().

This change adds the missing check and a regression test for passing
':'. It also simplifies the code and tightens up an out-of-bounds check
to make it off-by-one rather than off-by-two.