[observer] Improve performance for many links or monitors
When a process has many links and/or monitors, it could earlier take
very long time to display the process information window. This is now
improved by only showing a few links and monitors, and then an link
named "more..." to expand the rest.

Reading of the "Link list" from a crashdump is also improved.