Fix incorrect handling of floating point instructions
1a029efd1ad47f started to run the beam_block pass a second time.

Since it is run after introduction of the optimized floating point
instructions, it must handle those instructions correctly.
In particular, it must be careful when hoisting allocation
instructions. For example, the following code:

    {test_heap,{alloc,[{words,0},{floats,1}]},5}.
       .
       .
       .
    {fmove,{fr,2},{x,0}}.
    {allocate_zero,1,4}.

must not be rewritten to:

    {test_heap,{alloc,[{words,0},{floats,1}]},5}.
       .
       .
       .
    {allocate_zero,1,4}.
    {fmove,{fr,2},{x,0}}.

because beam_validator will not consider it safe. (The code may
actually be safe depending on what the code between the two allocation
instructions do.)

https://bugs.erlang.org/browse/ERL-555