beam_block: Combine blocks when running beam_block the second time
1a029efd1ad47f started to run the beam_block pass a second time,
but it did not attempt to combine adjacent blocks.

Combining adjacent blocks leads to many more opportunities for
optimizations.

After doing some diffing in generated code, it turns out that
there is no benefit for beam_split to split out line instructions
from blocks. It seems that the only reason it was done was to
slightly simplify the implementation of the no_line_info option
in beam_clean.