Fix size of fmt_double()'s format_str[] buffer
fmt_double() may write up to 8 characters into its format_str[]
buffer, which however only has room for 7 characters.

This case could be triggered by a call to erts_printf_format() with
any floating-point format that also includes #, and + or a space,
which may be uncommon, but a nif or driver could issue it.