Optimize non-strict equality check of binaries differing in size
This commit brings the perform of a non-strict equality check of
binaries into the same range as strict ones.

Currently, these checks can be significantly slower because they go
through the general comparison code which doesn't have an optimization
to check the size before comparing contents in the case that it's
checking equality.