Compute MD5s in dialyzer using the .beam file
The previous mechanism was based on the core file which
meant that for every module in the PLT, we had to fetch
its .beam file, retrieve Erlang AST, compile that down
to core, serialize it into a binary and then get its MD5.

In a project with stdlib, kernel, elixir and a small
application in the PLT,  relying only on the .beam sped
up --check_plt from 10s to 0.8s.