Expand ErlNifBinary by another word for future additions
When fixing OTP-14931/14932 we had to settle for a solution that
decreased performance in a few corner cases like multiple calls to
enif_make_binary, as the other options on the table required extra
fields which would have broken ABI compatibility.

This is ABI-compatible since the fields are unused and at the end.