beam_type: Make sure to kill all dependencies when killing a register
Make sure to kill all dependencies when a register is killed. For example,
in the following code, when the type information for {x,0} is killed in
the last instruction, there will still be type information for {x,1} referring
to {x,0}:

    {get_tuple_element,{x,0},0,{x,1}}.
    {test,is_eq_exact,{f,5},[{x,1},{atom,tag}]}.
    {get_tuple_element,{x,0},1,{x,2}}.
    {get_tuple_element,{x,0},2,{x,0}}.

This does not seems to have caused any problems in the past, but it may
cause problems in the future with a register allocator that reuses register
more aggressively.