Correctly handle get_map_elements with a literal map
A get_map_elements instruction that has a literal map operand
would never be translated to a i_get_map_element instruction.
That would be a problem for the following instruction:

    get_map_elements Fail #{} {x,0}, {x,1}

Since the key is not a literal, get_map_element must be used,
since get_map_elements requires that a hash value can be
calculated for each element.

When the instruction is translated to i_get_map_element, the
hash value will be set to 0 and an assertion will trigger in
the debug build.