Fix HiPE build for FreeBSD
FreeBSD used to define a MAP_NORESERVE that didn't do anything, but this
has been removed in current versions. While building HiPE on a FreeBSD
version without MAP_NORESERVE worked fine in 18.3.4, it fails in
configure for 20.2.3. Looking at the "maint" branch, it seems that
between these versions, commit 6c5277b1543 removed the requirement (that
apparently wasn't an actual requirement in 18) for MAP_NORESERVE on
FreeBSD in erts/emulator/sys/common/erl_mmap.h, and commit 03b00985c18
added the requirement in erts/configure.in. The change to configure.in
makes it consistent with erl_mmap.h.

Building on FreeBSD (with gcc) revealed another problem, not present
with HiPE enabled in 18 nor with HiPE disabled in 20: the compilation of
erts/emulator/drivers/unix/unix_efile.c fails due to a missing prototype
for finite(). The prototype isn't actually missing, but is hidden due to
the definition of _XOPEN_SOURCE in unix_efile.c (in 18 with HiPE
enabled, this generated a warning, but it causes failure due to the use
of -Werror=implicit in 20). To fix this, unix_efile.c was changed to
leave _XOPEN_SOURCE undefined also for FreeBSD.

As test case, build on e.g. FreeBSD 11.1 with gcc and --enable-hipe.