cerl_inline: Fix a name capture bug
The way variables created by make_template() are used, it is necessary
that the names are unique in the entire function.  This has not
happened to cause any problems in the past because all other compiler
passes created atom variable names, not integer variable names. If
other passes start to create integer variable names, this bug is
exposed.