Update types for posix error codes
I have read the man pages for most socket and file operations
on recent Linux, FreeBSD, OpenBSD and Solaris 10 and noted
the possible error codes.

Which error codes that are possible for file operations have
been updated in file:posix/0.  Error codes for socket operations
in inet:posix/0.  The latter refers to the former so it is
a superset, assuming that e.g sendfile and AF_UNIX socket
operations could cause socket operations to return any file
error code.  That is not entirely true, but could be,
especially in the future.

Added to file:posix/0 are:
  ebadmsg edeadlk edeadlock eftype emultihop enobufs enolck enolink
  enosr enostr enosys eopnotsupp eoverflow erange etxtbsy

Added to inet:posix/0 are all but:
  exbadport exbadseq file:posix()

These are still possible according to erl_posix_str.c,
but are not in file:posix/0 nor in inet:posix/0,
and many of them are not file nor inet related, but some might be:
  e2big eadv ealign ebade ebadfd ebadr ebadrpc ebadrqc
  ebadslt ebfont echild echrng ecomm edirty
  edom edotdot eduppkg eidrm einit eisnam
  elbin el2hlt el2nsync el3hlt el3rst
  elibacc elibbad elibexec elibmax elibscn elnrng
  enavail enet enoano enocsi enodata enoexec
  enonet enosym enotempty enotnam enotuniq
  eproclim eprocunavail eprogmismatch eprogunavail
  erefused eremchg eremdev eremote eremoteio
  eremoterelease erpcmismatch erremote eshutdown
  esrmnt esuccesss etime etoomanyrefs
  euclean eunatch eusers eversion exfull
  sysnotready vernotsupported ediscon enomore
  ecancelled einvalidproctable einvalidprovider eproviderfailedinit
  syscallfailure service_not_found type_not_found e_no_more
  e_cancelled unknown