beam_except: Enhance recognition of function_clause exceptions
Don't match exact BEAM instructions when trying to recognize
function_clause exceptions that should be replaced with a jump to the
func_info instruction. Instead, do a symbolic evaluation of the list
building code and see if the result is a list of the argument
registers.

While at it, also teach fix_block_1/2 to completely remove a test_heap
instruction before an exception generation instruction.