stdlib: Modify the printing of map associations with wWpP
Use the same depth for all (printed) elements of a map.

Since the order of keys can vary when printing a map--maps:iterator/1
and maps:next/1 are used--it is more consistent to print all
associations with the same depth.

If the associations printed are limited by the depth, the selection
of associations is arbitrary, as before.