[stdlib/sys]: install one debug function more times (#1781)
Allow installing multiple instances of sys debug function

This commit solves a bug which allowed installing {Fun,State} as sys debug function even if Fun was already installed. This happened in the case when the current State of the debug fun was undefined.

Also, the new format {Id,Fun,State} of debug functions can be installed, allowing multiple instances of the same fun.