erts: Let allocator pooled_tree also use Age Order
if configured for the allocator. This was not implemented
and pooled_tree always used address order first fit.

The "problem" was as a carrier can be part of both the pooled tree
(of its owner) and the allocation tree (of a foreign instance)
at the same time. Solved by duplicating 'birth_time' in both
AOFF_RBTree_t nodes for each carrier.

Blocks still cannot use age order and does not pay
any memory cost for birth_time.