v3_kernel: Remove guard optimizations
In 348b5e6bee2f, I added a guard optimization pass to v3_kernel.
At the time, Kernel Erlang was the intermediate representation
most suited for doing that kind of optimization. However, it
was not ideal from a maintenance standpoint, because none of
the current compiler maintainers have a deep grasp of Kernel
Erlang.

Therefore, since we now have SSA code, it is time to retire
the guard optimizations in v3_kernel.

This change will make it easier in the future to convert Core Erlang
directly to SSA code, eliminating Kernel Erlang.