Optimise beam_jump
This is an alternative to #1832.

The optimisation relies on special-casing the common pattern of
"renaming" a label by direct jump to another label. The change makes
beam_jump recognise couple more opportunities for optimisation.

The optimisation additionally avoids superfluous list concatenations by
only flattening the accumulator at the very end.