Make Linux OOM killer avoid killing heart process
When a Linux system runs out of memory, the Linux out-of-memory killer
selects a process to kill.  It then first kills the child processes of
that process, and if that didn't free enough memory, it kills the
process itself.

The Erlang "heart" program is run as a child process of the Erlang VM,
so if an Erlang VM is chosen, the OOM killer will first kill the heart
process and then the Erlang VM itself, defeating the purpose of having
a heart process.

This change, taken from port-linux.c in OpenSSH, attempts to change
the OOM score adjustment for the heart process, such that it will
never be killed by the OOM killer.  It then reverts this change when
restarting the Erlang VM, so that this invincibility doesn't get
passed on.

For this to work when running heart as a non-root user, the heart
binary needs to have the CAP_SYS_RESOURCE capability set:

    setcap CAP_SYS_RESOURCE+ep /path/to/heart