Move to a dirty scheduler even when we have pending system tasks
When a system task was enqueued on a process between being
scheduled and the check altered in this commit, we'd run dirty
code on a normal scheduler as the RUNNING_SYS flag wasn't set
and we wouldn't migrate back.

This change migrates us to a dirty scheduler instead, which will
immediately bounce us back to a normal scheduler where
RUNNING_SYS will be set appropriately.

This is caught fairly reliably by
process_SUITE:system_task_failed_enqueue on machines with a lot of
cores.