beam_type: Fix unsafe optimization
beam_type assumed that the operand for the bs_context_to_binary
instruction must be a binary. That is not correct;
bs_context_to_binary accepts anything. Based on the incorrect
assumption, beam_type would remove other test instructions.

The bug was introduced in eee8655788d2, which was supposed
to be just a refactoring commit.

https://bugs.erlang.org/browse/ERL-655